[#234] ♻️ Refactor : MyPageService -> UserService 로직 이동 리팩토링
